+++ app/src/main/java/riza/com/cto/core/Point.kt
@@ -3,6 +3,6 @@ package riza.com.cto.core
 import com.google.gson.annotations.SerializedName
 
 data class Point(
-    @SerializedName("lon") val x: Double,
-    @SerializedName("lat") val y: Double
+    @SerializedName("lon") val x: Float,
+    @SerializedName("lat") val y: Float
 )
